requirements.txt
absl-py==2.0.0
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.11
click==8.1.3
colorama==0.4.5
comm==0.1.2
cryptography==41.0.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
Django==4.1
docutils==0.18.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
Flask==3.0.0
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gplearn==0.4.2
grpcio==1.59.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.10.2
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.5.0
jupyter-server==1.23.5
jupyter-ydoc==0.2.2
jupyter_client==7.4.1
jupyter_core==5.2.0
jupyter_server_fileid==0.6.0
jupyter_server_proxy==4.0.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.19.0
keras==2.13.1
Kivy==2.0.0
kivy-deps.angle==0.3.1
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.0
pandoc==2.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
pipenv==2022.8.19
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
playsound==1.2.2
plumbum==1.8.1
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyee==8.2.2
pygame==2.1.0
Pygments==2.11.2
pypandoc==1.11
pypiwin32==223
pyppeteer==1.0.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
pywin32==303
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.0
simpervisor==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.1
uri-template==1.2.0
urllib3==1.26.8
virtualenv==20.16.3
virtualenv-clone==0.5.7
voila==0.4.0
vpython==7.6.4
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==3.0.1
widgetsnbextension==4.0.8
wrapt==1.15.0
y-py==0.5.5
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.8.0